refactor(module:date-picker): use directive instead of extends class
build: fix build
