+++ components/date-picker/lib/abstract-panel-header.ts
@@ -9,12 +9,12 @@
 import { EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges } from '@angular/core';
 import { CandyDate } from 'ng-zorro-antd/core';
 import { NzCalendarI18nInterface } from 'ng-zorro-antd/i18n';
-import { PREFIX_CLASS } from '../name';
-import { PanelMode } from '../standard-types';
 import { PanelSelector } from './interface';
 
+type PanelMode = 'decade' | 'year' | 'month' | 'week' | 'date' | 'time';
+
 export abstract class AbstractPanelHeader implements OnInit, OnChanges {
-  prefixCls: string = `${PREFIX_CLASS}-header`;
+  prefixCls: string = `ant-picker-header`;
   selectors: PanelSelector[] = [];
 
   @Input() value: CandyDate;
