+++ components/date-picker/month-picker.component.ts
@@ -6,44 +6,17 @@
  * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
  */
 
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component, forwardRef, Host, Input, Optional, ViewEncapsulation } from '@angular/core';
-import { NG_VALUE_ACCESSOR } from '@angular/forms';
+import { Directive, Host, Optional } from '@angular/core';
+import { NzDatePickerComponent } from './date-picker.component';
 
-import { NzNoAnimationDirective } from 'ng-zorro-antd/core';
-import { DateHelperService, NzI18nService } from 'ng-zorro-antd/i18n';
-import { AbstractPickerComponent } from './abstract-picker.component';
-import { DatePickerService } from './date-picker.service';
-import { PanelMode } from './standard-types';
-
-@Component({
-  encapsulation: ViewEncapsulation.None,
-  changeDetection: ChangeDetectionStrategy.OnPush,
+@Directive({
   selector: 'nz-month-picker',
-  exportAs: 'nzMonthPicker',
-  templateUrl: './abstract-picker.component.html',
-  host: {
-    '[class]': 'hostClassMap'
-  },
-  providers: [
-    DatePickerService,
-    {
-      provide: NG_VALUE_ACCESSOR,
-      multi: true,
-      useExisting: forwardRef(() => NzMonthPickerComponent)
-    }
-  ]
+  exportAs: 'nzMonthPicker'
 })
-export class NzMonthPickerComponent extends AbstractPickerComponent {
-  @Input() nzFormat: string = 'yyyy-MM';
-  @Input() nzMode: PanelMode | PanelMode[] = 'month';
-
-  constructor(
-    datePickerService: DatePickerService,
-    i18n: NzI18nService,
-    cdr: ChangeDetectorRef,
-    dateHelper: DateHelperService,
-    @Host() @Optional() public noAnimation?: NzNoAnimationDirective
-  ) {
-    super(datePickerService, i18n, cdr, dateHelper, noAnimation);
+// tslint:disable-next-line:directive-class-suffix
+export class NzMonthPickerComponent {
+  constructor(@Optional() @Host() public datePicker: NzDatePickerComponent) {
+    this.datePicker.nzMode = 'month';
+    this.datePicker.nzFormat = 'yyyy-MM';
   }
 }
