+++ package.json
@@ -110,6 +110,7 @@
     "shelljs": "^0.8.3",
     "sitemap": "^4.1.1",
     "ts-node": "~8.3.0",
+    "tsickle": "^0.38.1",
     "tslib": "^1.10.0",
     "tslint": "~5.18.0",
     "typescript": "~3.7.4",
